Login/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
pyotp==2.9.0
passlib
bcrypt==3.2.0
Flask-Mailman
itsdangerous
webauthn==2.3.0
cbor
Werkzeug==3.1.3
python-dotenv
